+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Changed
 
+- Load corpus timespan data in parallel when loading app [#437](https://github.com/spraakbanken/korp-frontend/issues/437)
 - More space in word picture tables [#102](https://github.com/spraakbanken/korp-frontend/issues/102)
 
 ### Fixed